Deployment/requirements.txt
azure-functions
openai
langchain
faiss-cpu
tiktoken
numpy